refactor(codegen): extract formControl into plugin folder
