+++ scripts/codegen/src/core/registry.ts
@@ -2,6 +2,7 @@ import { a11yPlugin } from "../plugins/a11y/index.ts";
 import { coveragePlugin } from "../plugins/coverage/index.ts";
 import { dependenciesPlugin } from "../plugins/dependencies/index.ts";
 import { examplesPlugin } from "../plugins/examples/index.ts";
+import { formControlPlugin } from "../plugins/formControl/index.ts";
 import { motionPlugin } from "../plugins/motion/index.ts";
 import { propsPlugin } from "../plugins/props/index.ts";
 import { tokensPlugin } from "../plugins/tokens/index.ts";
@@ -24,4 +25,5 @@ export const PLUGINS: readonly SubstratePlugin[] = [
   voidElementsPlugin,
   a11yPlugin,
   propsPlugin,
+  formControlPlugin,
 ];
