+++ scripts/codegen/src/plugins/formControl/index.ts
@@ -1,0 +1,10 @@
+import { z } from "zod";
+import type { SubstratePlugin } from "../../core/plugin.ts";
+
+export const formControlPlugin: SubstratePlugin = {
+  name: "formControl",
+  schema: {
+    atomic: { formControl: z.boolean().optional() },
+  },
+  emit: {},
+};
